* Commited Mikael Magnussons x86_64 compile fix for libmikey.
git-svn-id: svn://svn.minisip.org/minisip/trunk@3418 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmikey/mikey/MikeyPayloadCERT.cxx
@@ -45,7 +45,7 @@ MikeyPayloadCERT::MikeyPayloadCERT( int type, MRef<Certificate *> cert ){
 	this->type = type;
 	this->certLengthValue = cert->getDerLength();
 	this->certDataPtr = new byte_t[ certLengthValue ];
-	size_t size = certLengthValue;
+	unsigned int size = certLengthValue;
 	cert->getDer( this->certDataPtr, &size );
 }
